centralize target region utilities
